moved FileHelper into com.revature.util package
