+++ src/main/java/com/revature/services/S3StorageServiceImpl.java
@@ -16,7 +16,7 @@ import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.revature.helpers.FileHelper;
+import com.revature.util.FileHelper;
 
 @Service
 @Profile("!local")
